+++ mobile/PriceRecommender/app/src/main/java/com/example/pricerecommender/data/repositoryInterface/IProductRepository.kt
@@ -9,6 +9,6 @@ interface IProductRepository {
         userId: String,
         addressLat: Double,
         addressLng: Double,
-        range: Int
+        range: Float
     ): List<BestResult>
 }
